Revert "Fix PHP 7.1 Notice: A non well formed numeric value"
This reverts commit fc9d5c10e27136dad7865807c584db6471cfb658.

+++ DependencyInjection/OneupUploaderExtension.php
@@ -292,7 +292,7 @@ class OneupUploaderExtension extends Extension
     protected function getValueInBytes($input)
     {
         // see: http://www.php.net/manual/en/function.ini-get.php
-        $input = (float) trim($input);
+        $input = trim($input);
         $last  = strtolower($input[strlen($input) - 1]);
         $numericInput = substr($input, 0, -1);
